* Updated documentation.   - Added a 'modules' section to describe standalone models.  Currently, INFLOW     Is the only entry.   - Fixed variables in the doxyfile so that example files were found * Added two new sedflux utility functions:   - sedflux-make-bathy    : create a bathymetry file for sedflux-3D   - sedflux-make-sequence : create a sequence file for sedflux-3D * Fixed INFOW model.   - Created a better interface between INFLOW and sedflux.  This is now part of the     inflow library.   - Updated input files to be kvf, and csv format   - Updated command line options   - INFLOW now returns erosion and deposition rates separately. * Updated AVULSION model.   - Created a better UI for standalone avulsion model.   - Created an interface between AVULSION and sedflux.  The details are now kept in the     avulsion library.   - Rewrote the avulsion routine so that the reflection boundary condition is better     handled.  The domain is now scaled up to [-PI PI), the avulsion done, and the new     angle returned in the original domain. * Created the new object, Sed_riv.  Sed_riv replaces the public structure, Sed_river, with   a couple of changes:   - Sed_riv is completely opaque   - The avulsion data are now handled by the avulsion library rather than within     the Sed_riv data structure.  User data is held by a Sed_riv as a GDataList and     accessed through a key.  For instance, one can access the avulsion data with the     key "AVULSION".   - A Sed_riv can bifurcate any number of times, each time generating two new Sed_riv     children.  The children will have the same properties as the parent except the     discharge is split equally between the two.  This is done by equally dividing the     river width of the parent among the children. * Updated LIBUTILS library.   - new eh_dbl_array functions, including eh_dbl_array_rebin to rebin an array to a     coarser resolution while maintaining mass balance.   - Eh_key_file_entry data structure to describe a key-file template.  This is similar     to a GOptionEntry.   - eh_fopen_error function opens a file and sets an appropriate GError upon failure. * Upaded LIBSED library.  The big change here is the Sed_riv data structure described above.

+++ sedflux/src/storms.h
@@ -28,19 +28,14 @@ typedef struct
 {
    gboolean initialized;
 
-//   double average_storm;
-//   double variance;
-//   double storm_length;
    double last_k;
    double last_time;
    double alpha, beta;
    double fraction;   
    gboolean average_non_events;
-//   unsigned int seed;
-//   char state[256];
    GRand *rand;
    Eh_input_val wave_height;
-   long rand_seed;
+   guint32 rand_seed;
 }
 Storm_t;
